Tuesday Work
I've added a menu button that when pressed will display the menu, and when pressed again, the menu will hide just like a regular in game menu. I've also changed some of the dialogue to better fit the new format of the dialogue box. I've even made edits to the interaction path so that it would work better. There was an error that increased the affinity too much, but I've fixed that.