Added the Main Path and Code Comments
I added the Main Path system into the game. However, since the main path is new, what I created can be the basis of a more descriptive main path in the future. I also commented on my code for future coders to understand what I did. I tried to set up the tutorial, but without the artwork, it couldn't do a lot.